feat: added the category filter
